refactor(execution): extract submission outcome contract from order_submission
order_submission.py defined the OrderSubmissionOutcome result contract (and its
status enum) alongside the OrderSubmitter machinery. Move the inert outcome
types into execution/submission_outcome.py and re-export, so the 4 modules that
consume the outcome shape don't pull in the submitter.

Only inert result types moved; all submission logic stays in OrderSubmitter.
Re-export keeps identity stable; 365 execution tests pass.
